Add a base template and "All contacts view"
The templates were refactored to use a base
template to set the layout of the site.
The logged in user can now view all their
contacts.
Contacts currently can only be added through the
admin interface. The next stage is to add the
functionality to allow users to add contacts to
their lists.